Solution for servers using esx_multicharacter
Added a fix to prevent player coordinates error when connecting. multicharacter is not recognized by the player immediately and this causes this error.

SCRIPT ERROR: @dream-postal/client.lua:755: attempt to perform arithmetich on a nil value (upvalue 'POSTAL_BOSS_COORDS')
> fn (@dream-postal/client.lua:755)

You can improve it and adapt it to your programming style, I just added it to offer a solution!